# [Merged by Bors] - perf: improve performance of `to_additive`